Server/requirements.txt
asgiref
Django
django-ckeditor
django-cors-headers
django-js-asset
djangorestframework
pillow
sqlparse
tzdata
mysqlclient
whitenoise
requests